move footer and header into respective folders
